Create an example for scene debugging
